+++ app/Http/Controllers/Admin/MateriController.php
@@ -165,18 +165,16 @@ class MateriController extends Controller
         $picture = $dtMateri->picture;
         $file_path = public_path().'/uploads/materi/'.$picture;
         
+        //--Delete file di DB:
         $updateData['picture'] = null;
-        if(file_exists($file_path)){
-            // dd($file_path);
-            unlink($file_path);
-            if ($dtMateri->update($updateData)) {
-                return redirect()->back()->with('success', 'File gambar berhasil di hapus.');
-            }else{
-                return redirect()->back()->with('error', 'Error pada saat hapus file gambar. Silahkan hubungi Administrator.');
-            }
-       }else{
-        $dtMateri->update($updateData);
-        return redirect()->back()->with('error', 'File tidak ada');
-       }
+        if ($dtMateri->update($updateData)) {
+            //--Delete file di Path:
+            // ($thisVar != $thatVar ?: doThis()); //--1 line IF without else
+            // if ($thisVar == $thatVar) doThis();
+            if(file_exists($file_path)) unlink($file_path);           
+            return redirect()->back()->with('success', 'File gambar berhasil di hapus.');
+        }else{
+            return redirect()->back()->with('error', 'Error pada saat hapus file gambar. Silahkan hubungi Administrator.');
+        }
     }
 }
